Move the limiter into its own module
