+++ lib/Dialect/ESI/runtime/python/CMakeLists.txt
@@ -16,6 +16,7 @@ set(ESIPythonRuntimeSources
 
   esiaccel/components/__init__.py
   esiaccel/components/channel_arbiter.py
+  esiaccel/components/flow.py
 
   esiaccel/cosim/questa.py
   esiaccel/cosim/pytest.py
